Fix schema validation error: Change estado_id from SMALLINT to INTEGER
- Changed estado_id column type in cidades table from SMALLINT to INTEGER
- Fixes type mismatch with JPA entity mapping
- Resolves Hibernate schema validation error

+++ src/main/resources/db/migration/V10__create_location_cache.sql
@@ -14,7 +14,7 @@ CREATE INDEX idx_estados_sigla ON estados(sigla);
 -- Create cidades (Brazilian cities) cache table
 CREATE TABLE cidades (
     id SERIAL PRIMARY KEY,
-    estado_id SMALLINT NOT NULL REFERENCES estados(id),
+    estado_id INTEGER NOT NULL REFERENCES estados(id),
     nome VARCHAR(100) NOT NULL,
     codigo_ibge INTEGER NOT NULL UNIQUE,
     created_at TIMESTAMP NOT NULL DEFAULT CURRENT_TIMESTAMP
